Add setting to deal with transcoders that fudge PTS
Add a flag in settings (reportsUnreliablePTS) to disable the PTS scan
and create a default program_date value if faced with transcoders that
do not provider reliable PTS info.

On the fly transcoders can zero PTS when switching to another playlist
or seeking outside of the current set of transcoded segments.