Fixes bug 454506
Issue was when we parse type 4 SEI data (registered user data) it
partially parses a DTG1, AFD data without skipping past the bit it does
not parse.

We care nothing about the AFD data, but this change causes us to skip
it properly so the next user_data (the 708 CC) is not skipped.